Move mongo_helper.py to database module, refactor usage cases
