+++ database/mongo_helper.py
